fix: 데이터 타입 변경

+++ models.py
@@ -94,8 +94,8 @@ class Bcgdata(Base):
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     sequenceId = Column(Integer, ForeignKey('sequence.id')) 
     measureTime = Column(DateTime(timezone=True), nullable=False)
-    heart = Column(Integer, nullable=False)
-    respiration = Column(Integer, nullable=False)
+    heart = Column(Float, nullable=False)
+    respiration = Column(Float, nullable=False)
     
     sequence = relationship('Sequence', back_populates='bcgdatas')
